# C++: Remove the primitive field flow in FlowVar