This PR takes the training wheels off the AST-based field flow. The data flow library has always supported flow through fields by just treating a field as any other `Variable`, ignoring the qualifiers of field accesses. This has given some false flow, but it's been better than having no field flow at all. Now that proper inter-procedural flow through fields is working, we can remove the primitive intra-procedural version.

This change recovers the performance that was lost in #1715. The `getAReachedBlockVarSBB#ff` predicate run time drops from 21.4s to 8.3s on MongoDB and from 54.9s to 12.6s on MySQL.

I tried removing the `largeVariable` predicate that I introduced in #1715, but that caused MongoDB to take 29.8s, so it doesn't seem like a good idea.