@geoffw0 2019-08-19
I can confirm that performance is back to roughly pre-#1715 levels (I also saw no changes in results on these runs).
```
OverflowDestination.ql, LossyFunctionResultCast.ql, UnsafeDaclSecurityDescriptor.ql, SqlTainted.ql on httpd now takes 146s (was 157s), 7% faster.

OverflowDestination.ql, LossyFunctionResultCast.ql, UnsafeDaclSecurityDescriptor.ql, SqlTainted.ql on mysql now takes 575s (was 544s), 6% slower.

LossyFunctionResultCast.ql on mysql now takes 380s (was 346s), 10% slower.
```
I plan to have a slightly deeper look at this PR tomorrow.