@jbj 2019-08-20
The test revealed that we lost a good result because `localFlow` no longer attempts to track fields. @geoffw0, will this affect any security benchmarks? If not, I think we'll just have to live without this result until we can switch `localFlow` to be IR-based.