@jbj 2019-08-20
We could add the primitive field flow back to `localFlowStep` and have the global flow library use a version of `localFlowStep` that doesn't have primitive field flow. Then we'd avoid regressing functionality, but we'd regress performance.