@geoffw0 2019-08-20
> I plan to have a slightly deeper look at this PR tomorrow.

I now have, code and test changes LGTM.

> The test revealed that we lost a good result because localFlow no longer attempts to track fields. 

Is this specific to unions?  If so, I agree we can probably live with it.

> will this affect any security benchmarks?

`OverrunWrite.ql` is used in the samate benchmarks.  I've started https://jenkins.internal.semmle.com/job/Security/job/SAMATE/job/SAMATE-cpp/459/ to find out whether the change matters.