@jbj 2019-08-20
> Is this specific to unions? If so, I agree we can probably live with it.

No, it'll also happen with structs and classes. It's because `localFlow` now ignores fields completely. On current master it'll say there's flow from `foo.a = source()` to `sink(bar.a)` because it completely ignores field access qualifiers.