@geoffw0 2019-08-20
I see.  Why doesn't the inter-procedural flow doesn't apply in this case, because `mu` or `buffer` is a local???
```
{
  char buffer[25];
  ...
  mu->ptr = buffer;
  ...
  strcpy(mu->ptr, ...);
}
```