@jbj 2019-08-20
It’s because the underlying library uses `localFlowStep` directly and so
does not use inter-procedural flow at all. The use of data flow is buried
deep inside a library used in several places, and I would rather not change
that unless I have to. But we should probably do both the security
benchmark test and a dist compare for this PR.

tir. 20. aug. 2019 kl. 14.57 skrev Geoffrey White <notifications@github.com
>:

> I see. Why doesn't the inter-procedural flow doesn't apply in this case,
> because mu or buffer is a local???
>
> {
>   char buffer[25];
>   ...
>   mu->ptr = buffer;
>   ...
>   strcpy(mu->ptr, ...);
> }
>
> —
> You are receiving this because you authored the thread.
>
>
> Reply to this email directly, view it on GitHub
> <https://github.com/Semmle/ql/pull/1765?email_source=notifications&email_token=AACLIRKXZN7KNJR4GT6S5KDQFPS5HA5CNFSM4INAFGZ2YY3PNVWWK3TUL52HS4DFVREXG43VMVBW63LNMVXHJKTDN5WW2ZLOORPWSZGOD4WGBXQ#issuecomment-523002078>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AACLIRMH7T6RLSL4YBI46WLQFPS5HANCNFSM4INAFGZQ>
> .
>