@jbj 2019-08-21
There's one other thing I could try: calculating `localFlow` from a fifth copy of the data flow library where sources are expressions written to fields and sinks are field reads. I'll set up barriers so there's no interprocedural flow.