@geoffw0 2019-08-22
The samate tests passed, FWIW.  I'll review the other PR now.